[rosetta] Init operation_components.go & move respective code

[rosetta] Refactor formatTransaction for clarity


[rosetta] Move maxNumOfConstructionOps to operation_components.go

* Clarify unsupported type err msg


[rosetta] Implement framework for ConstructionPayloads

* Remove currying OperationComponents in metadata as PublicKeys is now
part of the request in the necessary endpoints
* Add respective checks for added PublicKeys
* Make getAddressFromPublicKey more general & update tests
